requirements.txt
torchvision == 0.22.1
torch == 2.7.1
numpy == 2.3.1
pandas == 2.2.3
pyyaml == 6.0.2
seaborn == 0.13.2
matplotlib == 3.10.0
profed == 0.7.3